refactor: Replace String buffer with Buffer type
This change further reduces string allocations by adding current line
length tracking to the buffer.  Previously, a separate string buffer was
allocated for current line length tracking in
format_function_declaration() to determine if the function defintion
should span multiple lines.  With the Buffer type, this is no longer
necessary.

This change also forces buffers to be allocated with an explicit
capacity.  This helps prevent excessive re-allocation as more content is
added to the buffer.  The main buffer is sized to be 1.5x the source
file content to allow some growth during transformation.