Replace final instance of symbol string matching
This required adding aliased symbols to the Symbol enum.  An alternative
to this would be to modify the grammar to remove unecessary hierarchy
for identifiers.  E.g. Replace all instances of 'class_identifier' with
'identifier'.  This could add some ambiguity to syntax trees due to loss
of information.  This loss of information could be addressed by adding
field information (see Tree-sitter grammar DSL).

This also required a bit of a hack for class declaration formatting.
Previous to this change, the ';' anonymous node was used to determine
the end of the class head (vs body).  The hack terminates when it sees
the first body node.  This too could be addressed with a grammar change
by putting the class head nodes under a class head node.  Currently they
are mixed in with class body items.